requirements.txt
numpy==1.14.4
PIL== 8.0.0
tensorflow-gpu==2.2.0
tensorflow_addons
keras==2.3.0